+++ tests/apps/demultiplex/conftest.py
@@ -347,3 +347,17 @@ def fixture_novaseq6000_flow_cell_sample_before_adapt_indexes() -> (
         SampleName="814206",
         Project="814206",
     )
+
+
+@pytest.fixture(name="novaseq6000_sample_sheet_path")
+def fixture_novaseq6000_sample_sheet_path() -> Path:
+    """Return the path to a NovaSeq S2 sample sheet."""
+    return Path(
+        "tests",
+        "fixtures",
+        "apps",
+        "sequencing_metrics_parser",
+        "Unaligned",
+        "Reports",
+        "SampleSheet.csv",
+    )
